reorginize elastic serach files
