+++ src/person/person.controller.ts
@@ -6,13 +6,13 @@ import { IPerson, IDomainUser, IDomainUserIdentifier } from './person.interface'
 import { IOrganizationGroup } from '../group/organizationGroup/organizationGroup.interface';
 import { OrganizationGroup } from '../group/organizationGroup/organizationGroup.controller';
 import { OrganizationGroupRepository } from '../group/organizationGroup/organizationGroup.repository';
-import { userFromString, getAllPossibleDomains, transformDomainUser, createDomainUserObject } from './person.utils';
+import { userFromString, getAllPossibleDomains, createDomainUserObject } from './person.utils';
 import * as utils from '../utils.js';
 import * as consts  from '../config/db-enums';
 import { PersonValidate } from './person.validate';
 import { search } from '../search/elasticsearch';
 import { config } from '../config/config';
-import { searchQuery } from './person.elastic.repository';
+import esRepository from './person.elastic.repository';
 
 export class Person {
   static _personRepository: PersonRepository = new PersonRepository();
@@ -293,49 +293,49 @@ export class Person {
    * Returns array of autocomplete suggestions on the "fullName" field
    * @param partialName the text to autocomplete, with minmum length of 2
    */
-  static async autocomplete(partialName: string) {
-    // check minimum length of 2
-    if (!partialName || partialName.trim().length < 2) {
-      return [];
-    }
-    // base match
-    const match_query = {
-      match: {
-        'fullName.autocomplete': {
-          query: partialName,
-        },
-      },
-    };
-    // allow fuzzy
-    const match_query_fuzzy = {
-      match: {
-        'fullName.autocomplete': {
-          query: partialName,
-          fuzziness: 'AUTO',
-        },
-      },
-    };
-    // search only for 'active' persons
-    const filter_active = {
-      term: { status: consts.STATUS[0] },
-    };
-    // construct the final query to send to ES
-    const query = {
-      query: {
-        bool: {
-          should: [
-            match_query, match_query_fuzzy,
-          ],
-          filter: filter_active,
-          minimum_should_match: 1, // necessary when using filter
-        },  
-      },
-    };
-    const results = await search<IPerson>(config.elasticSearch.personsIndexName, config.elasticSearch.defaultResultLimit, query);
-    return results.map(p => transformDomainUser(p));
-  }
+  // static async autocomplete(partialName: string) {
+  //   // check minimum length of 2
+  //   if (!partialName || partialName.trim().length < 2) {
+  //     return [];
+  //   }
+  //   // base match
+  //   const match_query = {
+  //     match: {
+  //       'fullName.autocomplete': {
+  //         query: partialName,
+  //       },
+  //     },
+  //   };
+  //   // allow fuzzy
+  //   const match_query_fuzzy = {
+  //     match: {
+  //       'fullName.autocomplete': {
+  //         query: partialName,
+  //         fuzziness: 'AUTO',
+  //       },
+  //     },
+  //   };
+  //   // search only for 'active' persons
+  //   const filter_active = {
+  //     term: { status: consts.STATUS[0] },
+  //   };
+  //   // construct the final query to send to ES
+  //   const query = {
+  //     query: {
+  //       bool: {
+  //         should: [
+  //           match_query, match_query_fuzzy,
+  //         ],
+  //         filter: filter_active,
+  //         minimum_should_match: 1, // necessary when using filter
+  //       },  
+  //     },
+  //   };
+  //   const results = await search<IPerson>(config.elasticSearch.personsIndexName, config.elasticSearch.defaultResultLimit, query);
+  //   return results.map(p => transformDomainUser(p));
+  // }
 
   static async searchPersons(query: object) {
-    return searchQuery(query);
+    return esRepository.search(query);
   }
 }
