+++ src/person/person.utils.ts
@@ -43,22 +43,5 @@ export function userFromString(uniqueID: string): IDomainUserIdentifier {
   return { name, domain };  
 }
 
-/**
- * Transform each domain user in the person to it's display form, 
- * returns new person object with the transformed domain users
- * @param person
- * @returns new person object with the transformed domain users
- */
-export function transformDomainUser(person: IPerson) {
-  const tPerson = { ...person };
-  if (!tPerson.domainUsers) return tPerson;
-  tPerson.domainUsers = (tPerson.domainUsers as IDomainUser[]).map((u) => {
-    const user: Partial<IDomainUser> = {};
-    user.uniqueID = `${u.name}${DomainSeperator}${u.domain}`;
-    domainMap.get(u.domain) && (user.adfsUID = `${u.name}${DomainSeperator}${domainMap.get(u.domain)}`);
-    user.dataSource = u.dataSource;
-    return user as IDomainUser;    
-  });
-  return tPerson;
-}
+
